ott-admin/backend/analytics_api/requirements.txt
aiohttp==3.8.1
fastapi==0.71.0
loguru==0.6.0
opensearch-py==2.0.0
pydantic==1.9.0
redis==3.5.3
redis-py-cluster==2.1.3
uvicorn==0.16.0
pytest==7.0.1
pandas
boto3
kinesis-python
geoip2==4.5.0